rename collision.ts to collisions.ts
