+++ src/collisions.ts
